requirements.txt
selenium~=4.16.0
Pillow~=10.0.1
pytesseract~=0.3.10
opencv-python~=4.8.1.78
wget~=3.2
